refactor: split the loop and version modules out of somewm.c
Moves WaylandSource, RefreshSource, some_glib_poll and the refresh cycle
into loop.c, and the --version helpers into version.c. Pure code motion.
