+++ version.h
@@ -1,0 +1,10 @@
+/*
+ * version.h - Version and environment info for --version
+ */
+#ifndef VERSION_H
+#define VERSION_H
+
+/* Print comprehensive version info in markdown format and exit. */
+void print_version_info(const char **paths, int num_paths);
+
+#endif /* VERSION_H */
